requirements.txt
fastapi==0.68.1
uvicorn==0.15.0
openai==0.27.0
slack_sdk==3.11.2
python-dotenv==0.19.1
pydantic==1.8.2
